Add multi-region support
Multi-region deployments aim to provide fault tolerance and reduce
the load on certain components such as databases or AMQP services.

Some of the Openstack services (e.g. Keystone, Horizon) will only
run inside the central region, comprised of nodes that have the
"region-controller" role.

The other nodes will connect to the central region using
cross-controller Juju relations.

To acommodate this scenario, we're making the following changes to
the Terraform modules:

* Added "is-region-controller" variable, skipping all Openstack API
  services other than Keystone and Horizon on region controller
  nodes
* Added "is-secondary-region" variable, used on secondary regions
  of multi-region environments. Keystone and Horizon will be excluded,
* Added "external-*-offer-url" variables, used to consume Keystone
  from a separate controller (i.e. the region controller)