Avoid schema changes
We've introduced a schema change by making various Openstack services
optional, using the "count" terraform meta-argument. It transforms
modules into tuples, which causes problems during snap upgrades.

To avoid this, we'll use the "scale" property instead, just like we
do for cinder-volume. This means that disabled applications will
always be deployed with the scale set to 0.