Adds Ironic to the terraform plan (#121)
Ironic is an OpenStack project which provisions bare metal machines.

- ironic-k8s charm contains the ironic-api and ironic-novncproxy services.
- nova-ironic-k8s charm contains the nova-compute service (configured
  with the Ironic driver), and the nova-novncproxy service.
- ironic-conductor-k8s charm contains the ironic-conductor service.

The ironic-k8s charm provides the ironic-api relation, which can be
added to the neutron-k8s charm, in which case the neutron-k8s charm
will starts its ironic-neutron-agent service.

The microceph's ceph-rgw-ready relation is required for ironic-conductor-k8s,
and it can be added to the glance-k8s charm.

Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>